Make symbols hashable
Using the name as basis for hashes may not be the best option,
because it will assign the same hash to Markers and UnknownSounds
and any other Symbol with name None. The other simple option, the
grapheme, is not very good, either. It may be necessary to think
of a more complex solution, after figuring out a few use cases.